move (Full)NodePrimitives to primitive-traits
